idk g

+++ src/core/physics/entity.rs
@@ -2,6 +2,6 @@ use struct_extension::extendable;
 
 #[extendable]
 struct PhysicsEntity {
-    position: (u32, u32),
-    velocity: (u32, u32)
+    position: (f64, f64),
+    velocity: (f64, f64)
 }
